Update dashboard.html
Changes made : removed js script cause i created a seprate file for it,
also the bootstrap links cause we arent using that at all.
added links to nav bar that send user to respective pages 
added to "the profile class"... basically added a menu bar that has the option to delete. 
A course also to change the theme display (dark or light mode).